refactor(migration): split the save-sort migration out of the home migration
Merging this branch onto main put services/migration.py at 1032 lines against
the gate's 1000 threshold. Neither side crossed it alone: main added the
realpath seams and this branch added the save answer, and the gate refuses an
allowlist entry for a module that crosses the line, by design.

The seam is the one the module already had. The two migrations share nothing
but the mechanics of moving a file to a path something may already occupy: they
have different triggers, different sources of truth for the destination, and
different conflict rules -- a strategy the user picks against newest-wins
resolved in place. A static analysis of the class confirmed it: the save-sort
half reaches outside its own set for exactly two methods, both of them that
shared move.

So the package is three parts. FileMover holds the move and the result tally.
SaveSortMigrator holds detection, the item collection, the emulator-answered
file set and the newest-wins resolution. MigrationService keeps the home
migration and stays the whole public surface, because callers reach one service
and refresh_state runs both halves.

Tests that reached into the moved private helpers now reach them on the
migrator, and the autouse loop fixture sets the migrator's loop too -- it took
its own copy of the seams at construction, so setting the service's alone left
it scheduling on the wrong loop.
